Add files via upload
In this project, I explore a weather data set from Szeged, Hungary, that contains hourly and daily summaries of various meteorological variables from 2006 to 2016. The data set includes information such as time, summary, precipType, temperature, apparentTemperature, humidity, windSpeed, windBearing, visibility, loudCover, and pressure.  My goal is to perform some exploratory data analysis (EDA), data preprocessing, and linear regression on this data set. I use Python and some popular libraries such as pandas, numpy, matplotlib, seaborn, and scikit-learn.
Some of the steps I take in this project are: 
Load and inspect the data set using pandas.
Check for missing values and outliers and handle them appropriately. 
Visualize the distribution and correlation of the variables using Matplotlib and Seaborn. 
Create some new features based on the existing ones. Split the data into training and testing sets. 
Build and evaluate a linear regression model to predict the temperature based on the other variables. 
Interpret the coefficients and the performance metrics of the model.